+++ Changelog.md
@@ -1,3 +1,4 @@
 # Changelog
 - Add trails to NEOs
 - Add gizmo to build cargo pods on large pod launcher
+- Add weighted choice when spawning asteroid ores